Got rid of slate-next and added mutations
