+++ examples/package.json
@@ -11,7 +11,6 @@
     "ory-editor-plugins-image": "0.0.0",
     "ory-editor-plugins-parallax-background": "0.0.0",
     "ory-editor-plugins-slate": "0.0.0",
-    "ory-editor-plugins-slate-next": "0.0.0",
     "ory-editor-plugins-spacer": "0.0.0",
     "ory-editor-plugins-video": "0.0.0",
     "ory-editor-renderer": "0.0.0",
