+++ packages/plugins/content/slate/package.json
@@ -23,16 +23,20 @@
   "dependencies": {
     "@material-ui/core": "^1.2.1",
     "@material-ui/icons": "^1.1.0",
-    "immutable": "^3.8.1",
+    "deep-rename-keys": "^0.2.1",
+    "immutable": "3.8.2",
     "ory-editor-core": "0.0.0",
     "ory-editor-ui": "0.0.0",
     "parse5": "^3.0.2",
     "ramda": "^0.24.1",
-    "react-portal": "^3.0.0",
+    "react-portal": "4.1.5",
     "redux-undo": "1.0.0-beta9-7",
     "selection-position": "^1.0.0",
-    "slate": "^0.21.4",
-    "slate-edit-blockquote": "^0.2.0",
-    "slate-edit-list": "^0.7.1"
+    "slate": "^0.37.3",
+    "slate-edit-blockquote": "^0.6.1",
+    "slate-edit-list": "^0.12.0",
+    "slate-html-serializer": "0.6.21",
+    "slate-plain-serializer": "^0.5.30",
+    "slate-react": "^0.15.4"
   }
 }
