+++ packages/plugins/content/slate/src/plugins/blockquote.js
@@ -52,15 +52,15 @@ export default class BlockquotePlugin extends Plugin {
         )
       )
 
-      let transform = editorState.transform()
+      let change = editorState.change()
 
       if (isActive) {
-        transform = transform.unwrapBlock(BLOCKQUOTE)
+        change = change.unwrapBlock(BLOCKQUOTE)
       } else {
-        transform = transform.wrapBlock(BLOCKQUOTE)
+        change = change.wrapBlock(BLOCKQUOTE)
       }
 
-      onChange(transform.apply())
+      onChange({ value: change.value })
     }
 
     const isActive = editorState.blocks.some(block =>
@@ -83,9 +83,12 @@ export default class BlockquotePlugin extends Plugin {
 
   name = 'blockquote'
 
-  nodes = {
-    [BLOCKQUOTE]: makeTagNode('blockquote')
+  schema = {
+    nodes: {
+      [BLOCKQUOTE]: makeTagNode('blockquote')
+    }
   }
+
 
   plugins = [
     createBlockquotePlugin({
@@ -100,7 +103,7 @@ export default class BlockquotePlugin extends Plugin {
     switch (el.tagName.toLowerCase()) {
       case 'blockquote':
         return {
-          kind: 'block',
+          object: 'block',
           type: BLOCKQUOTE,
           nodes: next(el.childNodes)
         }
@@ -108,10 +111,10 @@ export default class BlockquotePlugin extends Plugin {
   }
 
   serialize = (
-    object: { type: string, kind: string, data: any },
+    object: { type: string, object: string, data: any },
     children: any[]
   ) => {
-    if (object.kind !== 'block') {
+    if (object.object !== 'block') {
       return
     }
     switch (object.type) {
@@ -121,6 +124,18 @@ export default class BlockquotePlugin extends Plugin {
             {children}
           </blockquote>
         )
+    }
+  }
+
+  renderNode = props => {
+    switch (props.node.type) {
+      case BLOCKQUOTE: {
+        return (
+          <blockquote style={{ textAlign: props.node.data.get('align') }}>
+            {props.children}
+          </blockquote>
+        )
+      }
     }
   }
 }
