+++ packages/plugins/content/slate/src/plugins/code/index.css
@@ -1,4 +1,4 @@
-:not(pre)>code.ory-plugins-content-slate-next-code {
+:not(pre)>code.ory-plugins-content-slate-code {
     font-family: monospace;
     background-color: #eee;
     padding: 3px;
