+++ packages/plugins/content/slate/src/plugins/paragraph/index.js
@@ -31,13 +31,15 @@ export const P = 'PARAGRAPH/PARAGRAPH'
 export default class ParagraphPlugin extends Plugin {
   name = 'paragraph'
 
-  nodes = { [P]: Paragraph }
+  schema = {
+    nodes: { [P]: Paragraph }
+  }
 
   deserialize = (el: any, next: any) => {
     switch (el.tagName.toLowerCase()) {
       case 'p':
         return {
-          kind: 'block',
+          object: 'block',
           type: P,
           nodes: next(el.childNodes)
           // data: Data.create({ textAlign: el.attr('styles')['text-align'] })
@@ -46,15 +48,25 @@ export default class ParagraphPlugin extends Plugin {
   }
 
   serialize = (
-    object: { type: string, kind: string, data: any },
+    object: { type: string, object: string, data: any },
     children: any[]
   ) => {
-    if (object.kind !== 'block') {
+    if (object.object !== 'block') {
       return
     }
     switch (object.type) {
       case P:
         return <p style={{ textAlign: object.data.get('align') }}>{children}</p>
+    }
+  }
+
+  renderNode = props => {
+    switch (props.node.type) {
+      case P: {
+        return (
+          <Paragraph {...props} />
+        )
+      }
     }
   }
 }
